requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
lightgbm==4.3.0
xgboost==2.0.3
optuna==3.6.1
statsmodels==0.14.2
holidays==0.34
matplotlib==3.8.4
seaborn==0.13.2
pytest==8.2.2
python-dotenv==1.0.1
typer==0.12.3
mlflow
